Workout Page commit
Workout page as a standalone app. You can run it by calling the
MainApp.java.

Features implemented:
Reads from an xml file that contains the data for individual WORKOUTS.
These workouts are stored in an ObservableList and displayed in a table
which is placed on the left side of a Split pane. 

Create New Workout button takes you to a dialog which prepares the
workout metadata: the user can set a name and author. If the user
completes this form correctly, the newly created workout is added to the
ObservableList of workouts.

You can also save the workout library. (created workouts are saved too).

The application saves the previously loaded xml FILE PATH with
Preferences. This means that you have to load the xml once, the first
time you load the app. Need to work on automatically loading the correct
xml library of workouts on startup.

The other side of the split pane displays additional information for the
SELECTED workout, so that the user has more of an idea of what the
workout entails.


FXML:
The GUI itself was designed in Scene Builder which creates fxml files.
These are placed in a view package, which also contains controllers
which enable the MainApp to initialise the layouts.  

This is all contained in a BorderPane. South there is a progress bar and
labels for level and xp, and there is space for the character's avatar
on the left.