Added account package with all the classes related to the server-client communication. Also added resource folders for clientAccount files and serverAccount files.
It's functionality so far is: create account, login new, login, save
all, save selected and logout. 

To test, run the ServerSide class first and wait for the mouse loading
icon to disappear before running the communication tests. The order of
the messages sent is very important, as you need to make sure an account
is created before you try to save or login to it. You will also have to
close each thread manually at the end as I haven't implemented thread
merging yet.

Incomplete classes are: Communicate and ServerThreadManager.
These will be used to manage the separate server threads and pass friend
information to each user. Also a potential minigame will run over this
communication similar to that of pokemon.