Account Package
Added the account package containing the classes associated with
server-client communication.
Added resource folders for the client and server accounts.

DO NOT INTEGRATE WITH REST OF CODE UNTIL I HAVE COMPLETED THE
SERVERTHREADMANAGER CLASS.

Only basic functionality has been implemented so far including: create
new account, login new, login, save all, save selected and logout.

To test, run the ServerSide class and wait for the mouse loading icon to
finish before running the CommunicationTest class. The order in which
you send the messages from CommunicationTest is very important. For
example an account needs to be created before you can save or login to
it.

Incomplete Classes: Communicate and ServerThreadManager. These will be
use3d to send information to users about their friends activity and also
potentially used to play a minigame similar to that of pokemon.