NEW PWS UPDATE
I have modified the pws after the meeting with all the PMs and this is
the final version we will be using. For our personal copy, I have left
our optional attributes relating to workouts.

NB - I don't know why it's made me upload all the photos that were
originally in the project, but hopefully it hasn't changed anything