figured out how to do screen changes
the problem was that i was trying to set the screen from a java class that did not implement controllable.
every java screen/fxml controller needs to implement controllable.
the screenFlowController sets the its own StackPane as the parent for each screen.
screens are loaded into the hashmap when the app starts up and then the screenController.setScreen() is called on button handlers.