Diet Planner Updates Immediately and Workout Library shows XML Workouts
Adding a meal in Diet Menu is now displayed in Diet planner.
This was done by making a new DietPlanner in the ScreenController (which has been renamed to
StackPaneUpdater). The old DietPlanner is removed from the bank of screens and replaced by the new one.

Workout Library, renamed to Workout Browser, now looks for workouts stored in res/xml and displays
them in the table. The RHS updates when you select an other item in the table.